+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 - BREAKING: Rename `ShopifyAPI.CacheSupervisor` to `ShopifyAPI.Supervisor`.
+- Upgrade `AppServer`, `ShopServer`, and `AuthTokenServer` to use ets-backed caching.
 
 ## 0.9.4