+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - BREAKING: Rename `ShopifyAPI.CacheSupervisor` to `ShopifyAPI.Supervisor`.
 - Upgrade `AppServer`, `ShopServer`, and `AuthTokenServer` to use ets-backed caching.
+- Change default Shopify API version to `2020-10`.
 
 ## 0.9.4